feat(server/gateway): 라우터 경로 변경 및 불필요한 파일 삭제
더이상 passport 모듈 사용하지 않음.
auth 라우터들을 폴더 안에서 정리.
