+++ server/api-gateway/routes/auth/index.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const partnersRouter = require("./auth.partners");
-const usersRouter = require("./auth.users");
+const partnersRouter = require("./partners");
+const usersRouter = require("./users");
 
 router.use("/partners", partnersRouter);
 router.use("/users", usersRouter);
